This is a simple crud with JWT flow for practice

+++ server/src/services/todo.services.js
@@ -1,17 +1,21 @@
 import Todo from "../models/todo.model.js";
 
-export const createTodo = async (title) => {
-  return await Todo.create({ title });
+export const createTodo = async (todoData) => {
+  return await Todo.create(todoData);
 };
 
-export const getTodo = async () => {
-  return await Todo.find();
+export const getTodo = async (userId) => {
+  return await Todo.find({ userId });
 };
 
-export const updateTodo = async (id, data) => {
-  return await Todo.findByIdAndUpdate(id, data, { new: true });
+export const updateTodo = async (id, data, userId) => {
+  return await Todo.findOneAndUpdate(
+    { _id: id, userId },
+    data,
+    { new: true }
+  );
 };
 
-export const deleteTodo = async (id) => {
-  return await Todo.findOneAndDelete(id);
+export const deleteTodo = async (id, userId) => {
+  return await Todo.findOneAndDelete({ _id: id, userId });
 };
